+++ R/ichimoku.R
@@ -80,11 +80,11 @@
 #' TKR <- sample_ohlc_data
 #'
 #' cloud <- ichimoku(TKR)
-#' plot(cloud)
-#' print(cloud[101:110, ], plot = FALSE)
+#' cloud
 #'
 #' kumo <- ichimoku(TKR, ticker = "TKR Co.", periods = c(9, 26, 52), keep.data = TRUE)
 #' summary(kumo)
+#' print(kumo, plot = FALSE, n = 20)
 #' plot(kumo, theme = "solarized", type = "bar", custom = "volume")
 #'
 #' @rdname ichimoku